Properly split components into smaller ones
LBSD-980
